improving conda commands
The addition of channels like menpo and conda-forge drastically change the dependency chain, so it's better to have these configurations only added for the newly created environment. This is achieved by adding the `--env` option to the `conda config` command.
Also, after adding the `usgs-astrogeology` channel, it's not required in the `install` command.